added admin auth checks

+++ src/controllers/cars.js
@@ -23,36 +23,42 @@ router.get('/cars/:carId', (req, res) => {
 
 /** Route to add a new car. */
 router.post('/cars/', (req, res) => {
-  console.log(req.body);
-  const car = new Car(req.body);
-  car.save().then((carResult) => {
-    return res.json({car: carResult});
-  }).catch((err) => {
-    throw err;
-  });
+  if (req.user.admin) {
+    console.log(req.body);
+    const car = new Car(req.body);
+    car.save().then((carResult) => {
+      return res.json({car: carResult});
+    }).catch((err) => {
+      throw err;
+    });
+  }
 });
 
 /** Route to update an existing car. */
 router.put('/cars/:carId', (req, res) => {
-  Car.findByIdAndUpdate(req.params.carId, req.body).then((car) => {
-    return res.json({car});
-  }).catch((err) => {
-    throw err;
-  });
+  if (req.user.admin) {
+    Car.findByIdAndUpdate(req.params.carId, req.body).then((car) => {
+      return res.json({car});
+    }).catch((err) => {
+      throw err;
+    });
+  }
 });
 
 /** Route to delete a car. */
 router.delete('/cars/:carId', (req, res) => {
-  Car.findByIdAndDelete(req.params.carId).then(() => {
-    console.log(req.params.carId);
-    return res.json({
-      'message': 'Successfully deleted.',
-      '_id': req.params.carId,
-    });
-  })
-      .catch((err) => {
-        throw err.message;
+  if (req.user.admin) {
+    Car.findByIdAndDelete(req.params.carId).then(() => {
+      console.log(req.params.carId);
+      return res.json({
+        'message': 'Successfully deleted.',
+        '_id': req.params.carId,
       });
+    })
+        .catch((err) => {
+          throw err.message;
+        });
+  }
 });
 
 module.exports = router;
